backend/requirements.txt
# Modern LangChain v1 ecosystem (all packages aligned to v1.x)
langchain>=1.0.0
langchain-community>=0.4.0
langchain-text-splitters>=1.0.0

# Groq LLM (free tier - used for all agents)
langchain-groq>=0.2.0

# Gemini embeddings only (free tier - used for RAG)
langchain-google-genai>=4.0.0
google-genai>=1.0.0

# LangGraph v1 - agent orchestration
langgraph>=1.0.0

# ChromaDB
langchain-chroma>=1.0.0
chromadb>=1.0.0

# Web Search
tavily-python>=0.5.0

# Backend
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sse-starlette>=2.1.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.10.0
httpx>=0.27.0
